requirements.txt
requests>=2.32.3
pandas>=2.2.2
mysql-connector-python==9.0.0
python-dotenv==1.0.1
